requirements.txt
# Core YouTube API Libraries
google-api-python-client==2.108.0
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data Validation and Settings
pydantic==2.5.0

# Retry Logic
tenacity==8.2.3

# Environment Variables
python-dotenv==1.0.0

# Rich Terminal Output
rich==13.7.0

# Optional Development Libraries
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
pylint==3.0.3